Lagre detaljar frå køyringa. Gjenstår å populere felta

+++ apps/etterlatte-behandling/src/main/resources/db/migration/V137__utvide_omregningskjoering.sql
@@ -2,5 +2,5 @@ ALTER TABLE omregningskjoering ADD COLUMN beregning_beloep_foer bigint null defa
 ALTER TABLE omregningskjoering ADD COLUMN beregning_beloep_etter bigint null default null;
 ALTER TABLE omregningskjoering ADD COLUMN beregning_g_foer bigint null default null;
 ALTER TABLE omregningskjoering ADD COLUMN beregning_g_etter bigint null default null;
-ALTER TABLE omregningskjoering ADD COLUMN beregning_brukt_omregningsfaktor bigint null default null;
+ALTER TABLE omregningskjoering ADD COLUMN beregning_brukt_omregningsfaktor numeric null default null;
 ALTER TABLE omregningskjoering ADD COLUMN vedtak_beloep bigint null default null;
